feat: enhance ticket deletion with form handling and feedback

+++ src/app/(root)/(protected)/ticket/[ticketId]/actions/delete-ticket.ts
@@ -1,17 +1,31 @@
 'use server';
 
-import { revalidatePath } from 'next/cache';
 import { redirect } from 'next/navigation';
 
 import { getAuth } from '@/actions/get-auth';
+import { InitialActionsState } from '@/constants/initial-create-state';
 import { prisma } from '@/lib/prisma';
 import { formErrorHandler } from '@/utils/form-error-handler';
-import { signInPath, ticketsPath } from '@/utils/paths';
+import { signInPath } from '@/utils/paths';
 
-export const deleteTicket = async (ticketId: string) => {
+export const deleteTicket = async (
+  prevState: unknown,
+  formData: FormData,
+): Promise<InitialActionsState> => {
   const { user } = await getAuth();
 
   if (!user) redirect(signInPath());
+
+  const ticketId = formData.get('ticketId') as string;
+
+  if (!ticketId) {
+    return {
+      status: 'error',
+      message: 'Ticket ID não encontrado',
+      fieldErrors: undefined,
+      payload: undefined,
+    };
+  }
 
   try {
     await prisma.ticket.delete({
@@ -20,10 +34,14 @@ export const deleteTicket = async (ticketId: string) => {
         userId: user.id,
       },
     });
-  } catch (error) {
-    return formErrorHandler(error);
-  }
 
-  revalidatePath(ticketsPath());
-  redirect(ticketsPath());
+    return {
+      status: 'success',
+      message: 'Ticket excluído com sucesso',
+      fieldErrors: undefined,
+      payload: undefined,
+    };
+  } catch (error) {
+    return formErrorHandler(error, formData);
+  }
 };
